Move capabilities getter to libbeat
